Use working tree to restore icons
Moving the index.html file into the root broke the path references to
icons. It seems to be a common problem that isn't just an update to the
Webpack path. This  commit uses [this
strategy](https://musteresel.github.io/posts/2018/01/git-worktree-for-deploying.html)
of creating a git working tree on mater which contains the generated
files.